CC-1669: Set no offset when no offset was found for topic partition
The HDFS connector always uses an offset of 0 when resetting offsets and there is no output in HDFS for the requested topic partition. This results in the connector _always_ starting at the beginning, regardless of the `consumer.auto.offset.reset` setting.

This commit changes the behavior to not explicitly set an offset when the connector found no output in HDFS for the requested topic partition. As a result, the connector will rely upon the consumer to start where it has been configured, either the beginning if `consumer.auto.offset.reset=earliest` (which is what Connect explicitly sets by default) or at the most recent offset if `consumer.auto.offset.reset=latest` when the worker is explicitly configured this way.

Incidentally, this new behavior matches that of the S3 connector.