CC-5969: Pass extra connector configs to the partitioner
The HDFS connector currently passes all of its known connector properties to the partitioner, but extra properties not known to the connector are excluded. With this change, the connector will also pass any extra connector configuration properties to the partitioner.

Added a unit test to verify that a custom partitioner will see the extra properties. This new test failed before the `HdfsSinkConnectorConfig` code was changed, and now passes.