RecoveryHelper to speed up recovery after restart
This patch introduces a change in WAL and recovery process.
The idea is to avoid expensive scan of all table files and maintain a recovery record in the WAL instead. This record is written in the beginning of WAL log and it is not surrounded with 'begin and 'end' markers.
There are following situations possible:
a. There is no recovery record, there are normal records in WAL
b. There is no recovery record, no other records in WAL
c. There is a recovery record, there are normal records in WAL
d. There is a recovery record, no other records in WAL
Since recovery record is writted in the beginning, then it contains the latest offset only in a case when there is nothing else in the log, or other records are invalid(temp files are deleted). So in cases a,c and d recovery  process will pick the committed file from WAL with highest offset - either from recovery record or from normal records.
In case b when WAL log is empty or doesn't exist - latest offset will be discovered through full recoursive folder scan.