Adds flush.file.size to allow file size based rotation
Being able to control the size of the written files is extremely helpful. This change gets us in the vicinity of such a feature.
A file size guarantee would require a major rewrite, as the current logic is heavily based on partitions. The file size based rotation
is implemented in such a way, that we rotate all open files for a given `TopicPartitionWriter` instance, when any file has reached
the configured limit.

See: https://github.com/confluentinc/kafka-connect-hdfs/issues/365